Extract global variable, shader and material storage
